+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #3.1.35.128 - Test
 + In game sound support
++ Custom window background images(skins)
 
 #3.1.35.127
 #3.1.34.127 - Test